feat : 필터 보드 구분 및 디자인 변경으로 인한 레이아웃 변경
